style: add border to loading pg transparent icon
